scripts/requirements.txt
httpx
pyyaml